Store source zoom directly in bounds.csv
The bounds.csv width and height columns exist only so aggregation covering can derive each file's source zoom. Replace them with zoom, validate the new schema when reading catalogs, and use the stored value directly during covering.
